Rename _common to _pyvrp (#263)
