+++ docs/source/api/crossover.rst
@@ -5,7 +5,7 @@
 Crossover operators
 ===================
 
-The :mod:`pyvrp.crossover` module provides operators that are responsible for generating a new :class:`~pyvrp._common.Solution` offspring from two parent solutions.
+The :mod:`pyvrp.crossover` module provides operators that are responsible for generating a new :class:`~pyvrp._pyvrp.Solution` offspring from two parent solutions.
 
 .. automodule:: pyvrp.crossover.selective_route_exchange
