+++ pyvrp/diversity/_diversity.pyi
@@ -1,4 +1,4 @@
-from pyvrp._common import Solution
+from pyvrp._pyvrp import Solution
 
 def broken_pairs_distance(first: Solution, second: Solution) -> float:
     """
